seperate board & game
